Import NALD sent date for return logs (#1089)
https://eaflood.atlassian.net/browse/WATER-5018

WRLS is taking over management of the 'returns leg' from NALD. Contingent on that is the ability to continue reporting on returns.

Our work with RDP is behind the service changes, but we must switch soon to support the first quarterly return submissions in July. So, we will ensure we can generate the same reports by building ad hoc queries.

We've spotted that the existing BOXI reports include `send_date`, which the job does not currently import from NALD.

This change updates the `licence-returns-import` process to grab the sent date from NALD and apply it to the WRLS return log.

---

**IMPORTANT!**

This won't be a one-to-one mapping because of how the process was implemented (admittedly, as per requirements). When the billing & data team took over managing return submissions, they also adopted the 'return cycles' concept. So, a return log's start and end dates would align with a cycle, for example, 2008-04-01 to 2018-03-31 (winter and all year).

Later NALD form logs align with this, but older ones were generated on schedules determined by the area teams. The WRLS return log implementation cannot handle this, so the import generates return logs based on the NALD return format (return requirements), not NALD form logs.

Our recent changes to include the missing submission data match NALD return lines to WRLS return logs by date.

We have highlighted this because it is possible for a WRLS return log to be assigned submission data from multiple NALD form logs, each with its own `sent_date`.

**NALD**

- ID FL1 - 2002-01-01 to 2002-12-31
- ID FL2 - 2003-01-01 to 2003-12-31

**WRLS**

- ID RL1 - 2001-04-01 to 2002-03-31
  - NALD FL1
- ID RL2 - 2002-04-01 to 2003-03-31
  - NALD FL1
  - NALD FL2
- ID RL3 - 2003-04-01 to 2004-03-31
  - NALD FL2

Without re-engineering WRLS, the only option we have is to take the latest date where a WRLS return log matches to multiple NALD form logs.

This issue also applies to NALD received dates, and this solution is _exactly_ how the previous team resolved the problem, and is how WRLS received dates have been imported from the very start.