Reduce return cycle queries to just one
https://eaflood.atlassian.net/browse/WATER-5007
https://eaflood.atlassian.net/browse/WATER-5015

Return cycles are both a reference record, and a dynamic one. Every return log should link to a return cycle, and once created return cycles do not change.

But unlike charge categories or purposes, every year we have to dynamically add two new return cycles (summer and winter all year).

The legacy import code would check for each return log added, if there is a corresponding return cycle. If not it would create it.

When we [Refactored all the overnight jobs into one](https://github.com/DEFRA/water-abstraction-import/pull/1063) we replaced this with an UPSERT. This got the same result but reduced the number of queries to the DB by half.

However, there are still more than 500K return logs which means each night we're doing the same amount of queries.

We didn't worry too much about this kind of stuff at the time of the refactoring, we as we expected the return log import including the missing return submission data to be run just one or two times before being disabled. This is because WRLS would have taken over the returns leg from NALD.

With WLRS taking over from NALD, that also means the reporting solution it uses will take over. However, there is concern about how it will handle the extra data.

So, now we're being asked to switch on the new import early, and keep it running indefinitely to confirm if this is the case. Meantime, we can get things ready for WRLS taking over.

Fortunately, to simplify how WRLS has to handle this problem, we've opted to make sure the missing historic return cycles (starting from when the first licence starts) are added to the service. And we know by end of June WRLS _must_ have taken over from NALD, which means we also know it will have handled creating the two new return cycles.

Because of this, we can now just fetch _all_ the return cycles with a single query, and then look in those results as we process each return log to find its match. That's 500K plus hits on the DB reduced to 1!