Housekeeping - delete unused arg in controller
When testing spotted that a log message was not appearing. Tracked it down to the controller was still passing in a placeholder index arg, but we'd dropped that concept from `licence-data-import.js` and how it logged things.

This placeholder arg was being read as the `log` arg by each of the 'process' tasks hence they were not logging when called directly from the controller.