Housekeeping - Drop index from party CRM v2 import
Like the other imports we wanted to drop the concept of an index. Again, args are getting confused. We got lucky that the overnight jobs arg was interpreted as false as we would have seen a log entry for all 74K parties!

But to do this, we need to control how the service gets called, hence the need for a `processParty()` function.