Update persist-returns to use the new data source
When called from the over night job, it will have fetched the return cycles and have passed them through via `process.js`.

This is because the licence-returns-import process is called per licence, and we want to fetch the return cycles only once. But we've added a bit of logic to handle it being empty, so we can still support the process being called by the testing endpoint.