Bulk insert return submission lines
https://eaflood.atlassian.net/browse/WATER-5015

When we [Refactored all the overnight jobs into one](https://github.com/DEFRA/water-abstraction-import/pull/1063), we didn't dig too deep into the performance of the existing queries. We expected the return log import, including the missing return submission data, to be run just one or two times before being disabled. This is because WRLS would have taken over the returns leg from NALD.

When WRLS takes over, it also means a new reporting solution will take over (PowerBI). However, there is concern about how the ingest service built for it will handle the extra data. So, we're now being asked to switch on the new import early and keep it running indefinitely to confirm whether this is the case.

This means we're doing a second pass of the queries the import service is firing to see if we can reduce them. We know in AWS you have to take into CPU utilization and your burst balance (check out [AWS import performance tuning](https://github.com/DEFRA/water-abstraction-import/pull/1083) for more info). If we can reduce the number of queries we're firing, we're hoping not only will we speed up the import process, but that we'll reduce the risk of impacting the service, especially if users start thinking about annual billing.

We've managed to reduce the count by more than 500K with our [Reduce return cycle queries to just one](https://github.com/DEFRA/water-abstraction-import/pull/1091).

The next 'quick-win' is to bulk insert any missing return submission lines instead of doing them one-by-one.