Move old lines check into own module
Checking if there are old lines is a prerequisite to importing the return submissions. We started by copying the functions out of the existing licence-returns-import into our new one, but then remembered we had the same calls in extract-old-lines.

Time to put them into their own thing, and even add some tests!