Add process to the job's licence-data-import step
Also highlights a :facepalm moment for us!

When LicenceReturnsImportProcess was also handling replicating the submission data, we were checking if the old lines exist. This meant two queries to the DB (one to check if the table is there, then another to check if it is populated). But we did this for _every_ licence being processed!!

Doh!

We only need to check this once, so along with the benefits LicenceSubmissionsImportProcess is bringing, we save another 148K queries to the DB!