Split replicate returns into its own process (#1094)
https://eaflood.atlassian.net/browse/WATER-5015

When we [Refactored all the overnight jobs into one](https://github.com/DEFRA/water-abstraction-import/pull/1063), we didn't dig too deep into the performance of the existing queries. We expected the return log import, including the missing return submission data, to be run just once or twice before being disabled. This is because WRLS would have taken over the returns leg from NALD.

When WRLS takes over, it also means a new reporting solution (PowerBI) will take over. However, there is concern about how the ingest service built for it will handle the extra data. So, we're now being asked to switch on the new import early and keep it running indefinitely to confirm whether this is the case.

This means we're doing a second pass of the queries the import service is firing to see if we can reduce them. We know in AWS you have to take into account CPU utilisation and your burst balance (check out [AWS import performance tuning](https://github.com/DEFRA/water-abstraction-import/pull/1083) for more info). Suppose we can reduce the number of queries we're firing. In that case, we'd like to speed up the import process and reduce the risk of impacting the service, especially if users start considering annual billing.

With our [Reduce return cycle queries to just one](https://github.com/DEFRA/water-abstraction-import/pull/1091), we've reduced the count by more than 500 K.

We then updated the app to [Bulk insert return submission lines](https://github.com/DEFRA/water-abstraction-import/pull/1092). This meant that for each return log with missing data, we reduced the query count to 1 from a maximum of 365.

That gave us the idea of bulk inserting the return logs. This would reduce the 580K queries we are currently firing to just 74K (one per licence).

However, replicating a return log is tightly bound to persisting the return log itself. For each return log we fire a query that tells us if

- it exists
- that the submission data exists

To bulk upload the return logs, we need to break this coupling. This change does that. We move the replication of return logs into its own process. It will be called as part of the existing 'src/modules/import-job/lib/licence-data-import.js'. This means it will be passed a 'licence' instead of a return log.

So, we'll change the query it fires to find which of the licences' completed return logs are missing submission data. We'll then process them similarly to how we currently do it. However, instead of being based on the generated return log, we can use actual data from WRLS. Plus, we can update how we persist the submission data to use bulk inserts for both the lines and the versions.

The key is that after this change, we can return to the return log create/update process and switch it to use a bulk insert as well.