Update persist-returns to use both BULK and UPSERT
Using an UPSERT means we can avoid the query that checks whether we need to do an update or not. So, for example, if a licence generates a 100 return logs, that's 100 queries saved.

Next, we do what we came here for, which is to stop persisting the return logs one at a time and instead do them as a bulk insert. Again, if a licence generates a 100 return logs, that's 99 queries saved as we now do them in 1.