Add invalid-returns-clean process
https://eaflood.atlassian.net/browse/WATER-5015

When testing a change we added to In [Reduce return cycle queries to just one](https://github.com/DEFRA/water-abstraction-import/pull/1091) we exposed an issue with some return logs the legacy code was generating.

Essentially, if a licence or return version ends on the day a cycle starts (April 1 for winter and November 1 for summer), it was incorrectly assigning this as the end date for the last return log generated.

This lead to return logs being created as 2017-04-01 to 2018-04-01. This is invalid: a return log's period _cannot_ cross multiple return cycles.

[We fixed this](https://github.com/DEFRA/water-abstraction-import/pull/1093) (the PR explains the issue in more detail). We also know that the import will self heal. The correct return logs with valid dates will be created, and those with invalid dates will be marked as `VOID`.

The problem is, they'll still exist. We don't want any future functionality that assumes all return logs are 'valid' to break because we left these invalid return logs lying around, even if they are `VOID`.

So, we're adding a process to the import job to clean up these invalid return logs.

We need it to run _after_ our new import processes have cleaned things up. But once it's run once we can drop it from the import. It's work will be done! (we'll deal with that in a future release)