Fix not importing 0 qty NALD lines to WRLS
https://eaflood.atlassian.net/browse/WATER-5035

We've made a significant change to the overnight import process to include missing return submission data from NALD.

Billing & Data have been reviewing the results and have found an issue. It appears that where NALD has recorded a `0`, we are importing it as `NULL`. On the face of it, they may appear to mean the same thing, but they are, in fact, different.

`0` means the customer has explicitly told us that no water was abstracted at that time. `NULL` means no value has been submitted.

After investigating, we've spotted that the query which extracts the NALD lines for import is ignoring those with a quantity of `0`. This means because noting is matched to the generated WRLS line, we leave its quantity as `NULL`.

Simply removing the clause in the NALD lines queries resolves the issue!