Fix NALD received only rtns in WRLS for last cycle
https://eaflood.atlassian.net/browse/WATER-5055

In NALD, the form logs are generated at the start of the cycle. Like WRLS, you can apply a received date, but not add any quantities. However, in WRLS, we have a separate 'status' field, which is set to received at the same time. NALD doesn't have this.

When testing the historic NALD submission data in NALD, we found that we weren't handling NALD lines with zero and null quantities correctly. Essentially, we were ignoring them when we needed to include them.

But by doing so, it now cannot differentiate between

- A NALD form log where a user wants to record a received date
- A NALD form log that has been submitted, but where no values were provided

So, it has interpreted these records as 'nil returns' in WRLS and set them to `COMPLETE`.

We've agreed that NALD records with a received date and all null quantities should be interpreted as 'received only'. But when we checked there are more than 9K of them from as far back as 1994!

We're going to have a think about what we do with some of these very old returns. But in the meantime we'd like to clean up those in the last returns cycle at least (2024-04-01 onwards).

We need to come up with a query that 'corrects' those we've already imported as 'nil' returns. We then need to stop it happening to any more.

On that last point, we're not expecting them to. The ones they flagged in NALD as received only were done before the email reminders were generated on 1 April as a one-off exercise. Plus, we intend in the very near future to switch off this import.

So, we're going to 'hack' this! We won't touch the code that is generating the return submissions. It'll continue to create nil returns and mark the return log as `COMPLETE`. Instead, we'll just let the step that is needed to clean up those we've already imported handle any future ones create.

Simples!