Refactor /etl/returns endpoint
If the previous endpoint _had_ found 2000 returns, it would have involved making 2001 (one for the original 'return events' request) to water-abstraction-returns!

Now we do it in one. :-)

We know from speaking with the FME team that their tool is not really designed for talking to APIs, and supporting pagination of results was going to be a nightmare for them. So, they quite simply they don't!

We keep the structure of the response so as not to break their side, hence pagination is still included.