Refactor /etl/version/{id}/lines endpoint
This was the big one, and what most of the existing logic was related to, though we found it to be unnecessary.

Like with the other two endpoints, the legacy code was grabbing the return log and version data via the call to water-abstraction-returns. But because it all goes through hapi-pg-rest-api it had to make multiple calls and accept the data in its raw form. That meant it needed lots of logic to clean the data up.

It would then either generate some base nil lines, or fetch the data from the DB. It would then pass these into the final transformation layer, again applying logic to transform the data into what is expected in the response.

Using a daily full year return for example, this meant looping 365 times twice (!), all whilst applying various transformation logic to each row.

Just before that grab and then transform the return data, having started with grabbing and transforming the version data.

It is not having to go via water-abstraction-returns and hapi-pg-rest-api that explains why we can delete so much code. We do nearly all the transformation in the query within PostgreSQL, and can grab both the return log and version data in one hit (the pg connection we're stuck with meaning we have to handle converting any numbers in the results).

We've also refactored `generate-nil-lines.js`. Rather than generating a bunch of base line objects that then get iterated again by a transformation layer, we just do it in one.

Got a nil return? Then `generate-nil-lines.js` deals with the lines needed in the response. Got a completed response? Then we grab the rows and return them.

We still have to handle all the complications around how NALD handles 'weekly' returns (as in it doesn't really!) We also have to deal with only putting a 0 on nil lines that are within the abstraction period of the return log.

But we were able to do this for the most part without the need for extra logic.

One thing worth noting. Generally, return logs cover a full year, for example, 2024-04-01 to 2025-03-31. Daily and monthly returns fit in fine. Weekly however causes problems because the start and end week normally overlaps two periods. So, a week is assigned to a period based on whether it's Saturday falls within it.

The legacy `generate-nil-lines.js` had extra logic to stop generating lines that come after the last Saturday, but before the period ends.

However, it was unnecessary! The lines exist in NALD (it simply generates a line for each date between the return period start and end date) so its no issue if these get returned as FME will find a match in NALD.