Use SQL for API instead of calls plus reorg (#1104)
https://eaflood.atlassian.net/browse/WATER-5083

**Context**

With the RDP reporting solution not going to be ready in time for the transfer of the returns leg from NALD to WRLS, we are working on a number of contingencies.

One of the options we want to keep open is to not switch off the FME `WRLS->NALD` job.

**Technical background**

Currently, _all_ return logs in WRLS are generated by the water-abstraction-import based on NALD return requirement data. Licence by licence it

- generates 'return logs' (the return header record) based on the licence's return requirements (return formats in NALD terminology)
- attempts to insert them into WRLS
- If a 'conflict' is raised, it updates the existing record instead
- Any existing WRLS return logs not inserted or updated are marked as VOID, on the basis that the return requirements they were based on no longer exist

At the start of a return cycle, the B&D team runs a job in NALD that generates the return logs in NALD, using the same return requirement information. Therefore, although water-abstraction-import does not directly import NALD form logs as WRLS return logs, both systems will have a matching 'return log' at the time a submission is entered into WRLS.

When the FME job runs, it extracts submission data from the WRLS return log to apply to the NALD form log. It all works because the systems are in _sync_.

**Technical situation post-live**

At the point control switches from NALD to WRLS two key changes will come into effect.

- NALD will no longer manage the return requirements, nor generate form logs
- The import will 'switch off', and return logs will be created automatically in WRLS as changes are made to return requirements

There's more, but these two changes highlight that from this point on, the two systems will no longer be in sync.

FME cannot assume the return submissions it extracts from WRLS will match a form log in NALD.

**Idea to keep FME running**

We're aware that changes to return logs caused by changes to a licence's return requirements will require some additional consideration.

However, for those return logs that were _in sync_ at the time of the switch, changes to their return submissions should still be possible to extract by the FME job without issue. If so, it means we can keep the historic returns data in sync for some time yet, meaning BOXI can still be used as a 'source of truth' when it comes to return submission data.

**The change**

To avoid causing issues for FME, we need to exclude any return submissions (versions) that were not created via the import job from the results the API returns. Our problem is how the previous team engineered the current API. Rather than talk directly to the DB like the import jobs (even before our refactor most jobs they originally added did this) it instead makes calls to [water-abstraction-returns](https://github.com/DEFRA/water-abstraction-returns). It in turn is built on [hapi-pg-rest-api](https://github.com/DEFRA/hapi-pg-rest-api), a framework that 'simplified' things by automatically generating a [CRUD](https://en.wikipedia.org/wiki/Create,_read,_update_and_delete) based API for your data.

😱😭😩🤬

Seriously, it's the bane of our lives on WRLS, this being a classic example. We now need to filter one entity based on a property of another. We don't doubt you could do this with **hapi-pg-rest-api**, but it's not documented, and we have neither the time nor the will to go digging through the code to confirm if you can, then how.

Our FME friends have also informed us that they find the API slow. We know that when we refactored an import job to communicate directly with the database, we saw significant performance improvements.

So, we're killing two birds with one stone in this first change. We switch all the existing calls the API makes to **water-abstraction-returns** to use SQL queries instead. As is our way with things, we'll also use this opportunity to reorganise things to make them a bit clearer, especially as it doesn't look like the API is going anywhere soon!

We'll then tweak the version extraction process to filter out WRLS-sourced ones in a subsequent change.