Exclude WRLS return logs from FME API response
https://eaflood.atlassian.net/browse/WATER-5083

With the RDP reporting solution not going to be ready in time for the transfer of the returns leg from NALD to WRLS, we are working on a number of contingencies.

One of the options we want to keep open is to not switch off the FME `WRLS->NALD` job. For this to work though, when FME hits the `/etl/versions` endpoint it cannot be told about new return submissions (versions) linked to WRLS created return logs. For reasons we detail in WATER-5083 and the [change that preceded this one](https://github.com/DEFRA/water-abstraction-import/pull/1104), the FME job will break if returns that don't exist in NALD are included in the API response.

Now we've updated the API to no longer depend on calling [water-abstraction-returns](https://github.com/DEFRA/water-abstraction-returns) and its 'lock-in' to [https://github.com/DEFRA/hapi-pg-rest-api](hapi-pg-rest-api) we can easily tweak our queries to exclude WRLS sourced return logs and submissions from the responses.