Remove migrations from testing
If the previous team had created a separate database for unit tests we wouldn't have an issue. As it is, they use the same one for both.

Therefore, having dropped the schema the test migrations can no longer run because there is no migrations table.

So, we simply drop the concept of running the migrations as part of the unit tests.