Remove charge versions import process (#1111)
https://github.com/DEFRA/water-abstraction-team/issues/143

> See our Team issue for an expanded explanation

This change removes the legacy charge versions import process that, when manually triggered, would import NALD charge information into WRLS.

We believe it was created as a one-off when WRLS took over charging from NALD. However, it was retained as a means of obtaining 'realistic' data in a new development environment.

We don't want to maintain code that we are not using, and we shouldn't rely on such a process to obtain 'test data' in our service (and we haven't for a very long time!)

The other good news is that with this change, we can remove the `water_import` schema from our database.

The `water_import` schema served two purposes. When the project used `pgboss` message queues for orchestrating the import, it stored and archived the tables where jobs were held.

The other was as part of the process to import charge version information from NALD.

This project no longer uses `pgboss` or message queues. And in this change, we are removing the charge version import.

So we can say 'goodbye' to the `water_import` schema!